test(mri): relocate and refactor MRIExam and MRIExamId unit tests
